+++ packages/main/src/RPA/Desktop/Windows.py
@@ -11,8 +11,6 @@ import time
 from pathlib import Path
 from typing import Any, Optional
 
-from robot.libraries.BuiltIn import BuiltIn, RobotNotRunningError
-
 from RPA.Desktop import Desktop
 from RPA.Desktop.Clipboard import Clipboard
 from RPA.Desktop.OperatingSystem import OperatingSystem
@@ -1638,13 +1636,7 @@ class Windows(OperatingSystem):
         if region is not None:
             region = Region(*region)
 
-        try:
-            dirname = BuiltIn().get_variable_value("${OUTPUT_DIR}")
-        except (ModuleNotFoundError, RobotNotRunningError):
-            dirname = Path.cwd()
-
-        path = Path(dirname, "images", clean_filename(filename))
-        path = Desktop().take_screenshot(path=path, locator=region)
+        Desktop().take_screenshot(path=filename, locator=region)
 
     def _parse_element_attributes(self, element: dict) -> dict:
         """Return filtered element dictionary for an element.