+++ docs/source/releasenotes.rst
@@ -12,6 +12,10 @@ Release notes
 
 - Library **RPA.Excel.Application**: Fix for Windows given pywin32 dependency update
 
+- Library **RPA.Robocorp.WorkItems**: Keyword `Parse Work Item From Email` for
+  retrieving the input item dictionary payload from the sent e-mail JSON body which
+  triggered the process
+
 11.5.2
 ------